Use filescoped namespaces for classes.
